Update DateTests.swift
The script above is a set of test cases for a class that uses the Dependencies' library to test code that relies on the `Date` class. The `Dependency` property wrapper is used to create properties that can be overridden during the test.

The first test case, `testOverriding_Now`, uses the `withDependencies` function to set the `now` property of the` date` dependency to a specific date. The test then asserts that both the `now` property and the `date()` method return that same date.

The second test case, `testDateComparison`, creates two date objects and tests that one is less than the other.

The third test case, `testAddingTimeInterval`, creates a date object and adds a specific time interval to it, then asserts that the resulting date is the same as the date that would be created by the time interval since the reference date.

The fourth test case, `testChangingTimeZone`, creates a date object and changes its time zone to UTC and asserts that the date's time zone is UTC.


> All tests use the `withDependencies` function to set the `now` property of the `date` dependency to a specific date before running the test.