Avoid watchOS test crash
We've been using `DispatchQueue`'s `setSpecific` and `getSpecific` to
detect the main queue, but this doesn't seem to work in watchOS (or
maybe I've been holding it wrong). I think in our case we really only
care if we're on the main thread, so I'm hoping a `Thread.isMainThread`
is enough.

If this fix is appropriate we should apply it to other repos that rely
on this, as well, like SnapshotTesting.

Fixes #120