Remove insecure setters
It is currently possible to override `fireAndForget`, `assert`, and
`precondition`, but it is not safe to do so. You could accidentally
introduce bad code into a release.

Because these helpers are specifically for making test behavior
predictable, we should not allow custom behavior that can leak into a
live app.