Don't cache preview app entry point dependencies
We have a longstanding gotcha where we tell folks to avoid instantiating
their root model on their app entry point because any dependencies on
the root model may negatively affect any preview in their application.

This PR detects when a dependency is accessed in a SwiftUI preview app
entry point and prevents it from influencing the cache using the call
stack symbols available. While this isn't a silver bullet, and any async
work kicked off from the app entry point could still affect things
negatively, this should hopefully be mostly an improvement on the status
quo and maybe we won't have to refer folks to this gotcha as often in
the future.