Deprecate preview dependencies trait
Due to a quirk or bug or Xcode previews, dependency traits are executed
for every preview in a file, not just the one that's running. This can
cause dependencies from other previews to trample over the preview you
are viewing.

Because of this, let's deprecate this functionality in favor of good ole
fashioned `withDependencies`, with the option of using
`prepareDependencies`.