Amended: Fix OpenURL crash on macOS Sequoia
Further testing has revealed that the crash stems from the completion handler for `openURL`. This may be a bummer for use-cases that require it to check whether an app was available to handle a URL scheme.

An attempt was made to have the best of both worlds by returning `.systemAction` in the `OpenURLAction` handler, however this causes the completion handler to never return.

Therefore, the only real fix is to always return `.handled`, which makes it equivalent to the watchOS implementation that always yields `true`.