Re-deprecate preview traits
While #353 resolved a big issue with using a preview trait to override
dependencies, there are sadly other intermittent tooling issues that we
have no control over.

Preview traits are just too much of a black box for us to reliably hook
into, so let's not steer folks the wrong direction.